refactor: restructure project following Feature-Driven Design (FDD) architecture
