+++ frontend/src/app/(public)/auth/login/page.tsx
@@ -1,0 +1,5 @@
+import Login from '@/features/auth/features/login/page';
+
+export default function LoginPage() {
+  return <Login />;
+}
