+++ frontend/src/app/(public)/auth/register/page.tsx
@@ -1,0 +1,5 @@
+import Register from '@/features/auth/features/register/page';
+
+export default function LoginPage() {
+  return <Register />;
+}
