+++ frontend/src/features/dashboard/components/ui/PerformanceCard.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useMemo } from 'react';
-import { useAppSelector } from '@/redux/hooks';
-import formatCurrency from '@/utils/currency-formatter';
-import { ByDateTimeAlgorithm } from '@/utils/algorithms';
+import { useAppSelector } from '@/shared/store/hooks';
+import formatCurrency from '@/shared/utils/currency-formatter';
+import { ByDateTimeAlgorithm } from '@/shared/utils/algorithms';
 
 import { Box, Paper, Typography } from '@mui/material';
 import { SparkLineChart } from '@mui/x-charts';
