+++ frontend/src/shared/store/features/currency/currencySlice.ts
