+++ frontend/src/shared/styles/globals.css
