+++ frontend/src/shared/types/globals.d.ts
