+++ frontend/src/shared/utils/algorithms.ts
