+++ frontend/src/shared/utils/currency-formatter.ts
